backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
pandas
sqlalchemy
psycopg2-binary
alembic
httpx
rapidfuzz
fasttext